Ellipsoid Kernel Creation fixes

+++ src/c/Common/CHelpers.h
@@ -4,6 +4,6 @@
 #include <memory.h>
 #include <complex>
 
-double* createEllipsoidKernel(Vec<size_t> radii, Vec<size_t>& kernelDims);
+float* createEllipsoidKernel(Vec<size_t> radii, Vec<size_t>& kernelDims);
 
 int calcOtsuThreshold(const double* normHistogram, int numBins);
